refactor(og): generate lgtm's favicons via @kkhys/og
Generate the five favicons via @kkhys/og createFaviconGenerators (bound to lgtm's green gradient) and serve favicon/OG routes through the shared handlers, dropping the duplicated local generators and _util. The default OG card (bespoke stroked LGTM) and the per-id image pipeline stay app-local. OG PNG output is byte-identical.
