--- apps/lgtm/src/components/seo/favicon/apple-touch-icon-png.tsx
@@ -1,9 +1,0 @@
-import { generateIcon } from "./utils";
-
-export const AppleTouchIconPng = async () => {
-  return (await generateIcon(
-    { width: 180, height: 180 },
-    { width: 140, height: 140 },
-    "png",
-  )) as Buffer;
-};
