--- apps/lgtm/src/components/seo/favicon/icon-192-png.tsx
@@ -1,9 +1,0 @@
-import { generateIcon } from "./utils";
-
-export const Icon192Png = async () => {
-  return (await generateIcon(
-    { width: 192, height: 192 },
-    undefined,
-    "png",
-  )) as Buffer;
-};
