--- apps/lgtm/src/components/seo/favicon/icon-512-png.tsx
@@ -1,9 +1,0 @@
-import { generateIcon } from "./utils";
-
-export const Icon512Png = async () => {
-  return (await generateIcon(
-    { width: 512, height: 512 },
-    undefined,
-    "png",
-  )) as Buffer;
-};
