--- apps/lgtm/src/components/seo/favicon/icon-mask-png.tsx
@@ -1,9 +1,0 @@
-import { generateIcon } from "./utils";
-
-export const IconMaskPng = async () => {
-  return (await generateIcon(
-    { width: 512, height: 512 },
-    { width: 409, height: 409 },
-    "png",
-  )) as Buffer;
-};
