+++ apps/lgtm/src/pages/api/favicon/icon-512.png.ts
@@ -1,4 +1,4 @@
-import { Icon512Png } from "#/components/seo/favicon/icon-512-png";
-import { createPngHandler } from "./_util";
+import { createPngHandler } from "@kkhys/og/handlers";
+import { Icon512Png } from "#/components/seo/favicon";
 
 export const GET = createPngHandler(Icon512Png);
